Move HTTP related things into client.go
