+++ client.go
@@ -1,0 +1,53 @@
+package main
+
+import (
+	"fmt"
+	"io/ioutil"
+	"log"
+	"net/http"
+	"net/http/cookiejar"
+)
+
+const (
+	baseAuthTermURL = "https://ssb.cc.nd.edu/StudentRegistrationSsb/ssb/term/search?mode=search&term="
+)
+
+func setupClient() (*http.Client, error) {
+	cookieJar, err := cookiejar.New(nil)
+
+	if err != nil {
+		log.Fatal(err)
+		return nil, err
+	}
+
+	client := &http.Client{
+		Jar: cookieJar,
+	}
+
+	return client, err
+}
+
+func authenticateClient(c *http.Client, term string) {
+	authURL := baseAuthTermURL + term
+	fmt.Println(authURL)
+
+	_, err := doGet(c, authURL)
+	if err != nil {
+		log.Fatal(err)
+	}
+}
+
+func doGet(c *http.Client, URL string) (string, error) {
+	resp, err := c.Get(URL)
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	response, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+
+	return string(response), err
+}
